kWh more specific than integer
First proposed change (in GitHub), no knowlegde of Python, or the implications, but I'd like the kWh as a float so that utility_meter cost calculations are more precise than based on a rounded integer.

+++ homeassistant/components/plugwise/sensor.py
@@ -380,7 +380,7 @@ class PwPowerSensor(SmileSensor, Entity):
         if data.get(self._sensor) is not None:
             measurement = data[self._sensor]
             if self._unit_of_measurement == ENERGY_KILO_WATT_HOUR:
-                measurement = int(measurement / 1000)
+                measurement = float(measurement / 1000)
             self._state = measurement
             self._icon = CUSTOM_ICONS.get(self._sensor, self._icon)
